Move contents of `exit.c` to `runtime.c`
